Snapshot commit - initial import.
This ansible code is a prototype. It's been used to test how
to set up Apache2-based frontend nodes behind an HAproxy, which
server a Wordpress with mariadb backend.

The matching infrastructure can be created with the playbook
in the wut-infra directory, but you need to create a clouds.yaml
first.